moved some of the staticmethod functions from biofilm into new script utils.py. Refactoring of everything else to work with new python file
